Allow rooms to be visited directly
The store is first checked to see whether it contains room data with
the requested room Id (router param). If so, this data is loaded into
the page as the currentRoom. If not, a GET request is sent to load the
data from the backend. If the server does not have a room with that ID
either, a message is displayed indicating that the requested room does
not exist.
Issue #4