Split advanced info into a component
Also cleaned up the onEdit method. Made available in the RoomView
parent, although if I use this component somewhere else this will become
problematic. Will I need to use this component elsewhere -- does it
deserve to be a component? Definitely easier to parse what's going on
in RoomView in this format.
Issue #27