Add creature abilities, actions
Can't for the life of me figure out why I'm not able to name the
creature attribute creatureAbilities just plain abilities - I get an
error when I try to create a creature with this setup. Possibly some
level of spooky caching because I used to have what is now called
abilityScores as abilities, and its old type (embedded object) is
clashing with the type of this new thing?