Allow creature that is associated with a room to be deleted
The creature is first removed from all associated rooms, then it is
deleted.
I should probably emit a warning when this happens (e.g. when the user
clicks delete, first query to find a list of associated rooms and say
HEY ARE YOU SURE you will be deleting the creature from the following
rooms etc. Same deal for edit perhaps..?